+++ assembly/src/assembler/instruction/mod.rs
@@ -423,9 +423,11 @@ impl Assembler {
             },
 
             Instruction::Debug(options) => {
-                block_builder.push_decorator(Decorator::Debug(
-                    options.clone().try_into().expect("unresolved constant"),
-                ))?;
+                if self.in_debug_mode() {
+                    block_builder.push_decorator(Decorator::Debug(
+                        options.clone().try_into().expect("unresolved constant"),
+                    ))?;
+                }
             },
 
             // ----- emit instruction -------------------------------------------------------------