+++ miden/src/cli/run.rs
@@ -41,7 +41,7 @@ pub struct RunCmd {
     #[clap(short = 't', long = "tracing")]
     tracing: bool,
 
-    /// Enable tracing to monitor execution of the VM
+    /// Enable debug instructions
     #[clap(short = 'd', long = "debug")]
     debug: bool,
 }
@@ -104,31 +104,25 @@ impl RunCmd {
 // HELPER FUNCTIONS
 // ================================================================================================
 
-pub fn create_execution_options(cmd: &RunCmd) -> Result<ExecutionOptions, Report> {
-    let mut execution_options =
-        ExecutionOptions::new(Some(cmd.max_cycles), cmd.expected_cycles, cmd.tracing)
-            .into_diagnostic()?;
-
-    if cmd.debug {
-        execution_options = execution_options.with_debugging();
-    }
-
-    Ok(execution_options)
-}
-
 #[instrument(name = "run_program", skip_all)]
 fn run_program(params: &RunCmd) -> Result<(ExecutionTrace, [u8; 32]), Report> {
     // load libraries from files
     let libraries = Libraries::new(&params.library_paths)?;
 
     // load program from file and compile
-    let program =
-        ProgramFile::read(&params.assembly_file)?.compile(&Debug::Off, &libraries.libraries)?;
+    let program = ProgramFile::read(&params.assembly_file)?
+        .compile(&Debug::from(params.debug), &libraries.libraries)?;
 
     // load input data from file
     let input_data = InputFile::read(&params.input_file, &params.assembly_file)?;
 
-    let execution_options = create_execution_options(params)?;
+    let execution_options = ExecutionOptions::new(
+        Some(params.max_cycles),
+        params.expected_cycles,
+        params.tracing,
+        params.debug,
+    )
+    .into_diagnostic()?;
 
     // fetch the stack and program inputs from the arguments
     let stack_inputs = input_data.parse_stack_inputs().map_err(Report::msg)?;
@@ -142,78 +136,4 @@ fn run_program(params: &RunCmd) -> Result<(ExecutionTrace, [u8; 32]), Report> {
         .wrap_err("Failed to generate execution trace")?;
 
     Ok((trace, program_hash))
-}
-
-// TESTS
-// ================================================================================================
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_create_execution_options_with_default_settings() {
-        let cmd = RunCmd {
-            assembly_file: PathBuf::from("test.masm"),
-            expected_cycles: 64,
-            input_file: None,
-            library_paths: vec![],
-            max_cycles: 4294967295,
-            num_outputs: 16,
-            output_file: None,
-            tracing: false,
-            debug: false,
-        };
-
-        let result = create_execution_options(&cmd);
-        assert!(result.is_ok(), "Expected Ok, but got Err");
-
-        let options = result.unwrap();
-        assert_eq!(options.max_cycles(), cmd.max_cycles);
-        assert_eq!(options.expected_cycles(), cmd.expected_cycles);
-        assert!(!options.enable_tracing(), "Tracing should be disabled");
-        assert!(!options.enable_debugging(), "Debugging should be disabled");
-    }
-
-    #[test]
-    fn test_create_execution_options_with_tracing() {
-        let cmd = RunCmd {
-            assembly_file: PathBuf::from("test.masm"),
-            expected_cycles: 64,
-            input_file: None,
-            library_paths: vec![],
-            max_cycles: 4294967295,
-            num_outputs: 16,
-            output_file: None,
-            tracing: true,
-            debug: false,
-        };
-
-        let result = create_execution_options(&cmd);
-        assert!(result.is_ok(), "Expected Ok, but got Err");
-
-        let options = result.unwrap();
-        assert!(options.enable_tracing(), "Tracing should be enabled");
-    }
-
-    #[test]
-    fn test_create_execution_options_with_debugging() {
-        let cmd = RunCmd {
-            assembly_file: PathBuf::from("test.masm"),
-            expected_cycles: 64,
-            input_file: None,
-            library_paths: vec![],
-            max_cycles: 4294967295,
-            num_outputs: 16,
-            output_file: None,
-            tracing: false,
-            debug: true,
-        };
-
-        let result = create_execution_options(&cmd);
-        assert!(result.is_ok(), "Expected Ok, but got Err");
-
-        let options: ExecutionOptions = result.unwrap();
-        assert!(options.enable_debugging(), "Debugging should be enabled");
-    }
 }