+++ processor/src/lib.rs
@@ -181,7 +181,6 @@ where
     host: RefCell<H>,
     max_cycles: u32,
     enable_tracing: bool,
-    pub enable_debugging: bool,
 }
 
 #[cfg(any(test, feature = "testing"))]
@@ -197,7 +196,6 @@ where
     pub host: RefCell<H>,
     pub max_cycles: u32,
     pub enable_tracing: bool,
-    pub enable_debugging: bool,
 }
 
 impl<H> Process<H>
@@ -242,7 +240,6 @@ where
             host: RefCell::new(host),
             max_cycles: execution_options.max_cycles(),
             enable_tracing: execution_options.enable_tracing(),
-            enable_debugging: execution_options.enable_debugging(),
         }
     }
 
@@ -587,7 +584,7 @@ where
                 self.host.borrow_mut().set_advice(self, *injector)?;
             },
             Decorator::Debug(options) => {
-                if self.enable_debugging {
+                if self.decoder.in_debug_mode() {
                     self.host.borrow_mut().on_debug(self, options)?;
                 }
             },