+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Added `miden_core::utils::sync::racy_lock` module (#1463).
 - Updated `miden_core::utils` to re-export `std::sync::LazyLock` and `racy_lock::RacyLock as LazyLock` for std and no_std environments, respectively (#1463).
 - Made the undocumented behavior of the VM with regard to undefined behavior of u32 operations, stricter (#1480)
+- Debug instructions can be enabled in the cli `run` command using `--debug` flag
 
 #### Fixes
 
@@ -24,7 +25,6 @@
 #### Enhancements
 
 - Added `PartialEq`, `Eq`, `Serialize` and `Deserialize` to `AdviceMap` and `AdviceInputs` structs (#1494).
-
 
 ## 0.10.5 (2024-08-21)